update method from main 2

+++ src/main/java/com/project/credit/merchant/repository/MerchantRepository.java
@@ -3,5 +3,5 @@ package com.project.credit.merchant.repository;
 import com.project.credit.merchant.entity.Merchant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MerchantRepository extends JpaRepository<Merchant, Integer> {
+public interface MerchantRepository extends JpaRepository<Merchant, Long> {
 }
